Left Side Menu (#19)
Add a Left Side Menu with a quick add task feature to speed up the workflow of creating the most common tasks. 
Will make use of the screen estate.

In the future, the featured tasks should be determined by which tasks are most commonly created and completed by each user. 
-------------------------------------------------------------
Commits: 
---------------------------------------------------------------
* Allow the frontend localhost url through CORS
* Style changes -> ditch the grey
* Add sceleton for menu
* Change theme color for user
* Add ui for quick add
* Make header sticky
* Add quick add functionality
* Get quick add options from the backend
Simply returns a string array of titles to use. A smart way to determine what should come will come in the future
* Show the suggested quick adds
* Rename left menu title
* Format and fix linting issues